# api/openapi.yaml
openapi: 3.0.3
info:
  title: StudyWiki-Agent API
  version: 1.0.0
  description: |
    StudyWiki-Agent 服务契约 —— 基于 LangChain + LangGraph 的可增长本地 Wiki 知识库 API。

    统一约定:
    - 成功响应通常为 `{"status":"success","data":...}`（部分端点直接返回该字典，部分经 `ApiResponse` 包装并额外携带 `message`/`error_code`/`timestamp`）。
    - 错误响应统一为 `{"status":"error","error_code":"SW-...","message":"...","timestamp":"..."}`。
    - 除 `/health` 返回 `{"status":"ok","cards_count":N}` 外，业务接口顶层均含 `status` 字段。

    鉴权:
    - 当 `.env` 配置 `STUDYWIKI_AUTH_TOKEN` 时启用，业务 API 需携带 `Authorization: Bearer <token>`。
    - public 路径（无需鉴权）: `/`、`/health`、`/docs`、`/openapi.json`、`/api/openapi.yaml`、`/api/bootstrap/status`、`/api/bootstrap/test`、`/api/bootstrap/configure`，以及 `/static/*`。
    - 未授权返回 401，错误码 `SW-AUTH-001`。

    > 本文档为手写契约，与 `api/openapi.yaml` 一致；实际运行时的自动导出见 `/openapi.json`。
servers:
  - url: http://127.0.0.1:8000
    description: 本地开发服务器
...
paths:
...
  /api/quiz/generate/{card_id}:
    post:
      tags: [quiz]
      summary: 为卡片生成 3-5 道简答题
      parameters:
        - $ref: "#/components/parameters/CardId"
      responses:
        "200":
          description: 题目列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/QuizGenerateResult" }
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /api/quiz/grade:
    post:
      tags: [quiz]
      summary: AI 评分简答题
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuizSubmission" }
      responses:
        "200":
          description: 评分结果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/QuizGradeResult" }
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /api/quiz/merge/{card_id}:
    post:
      tags: [quiz]
      summary: 将 Quiz 问答与评分反馈融合回卡片
      parameters:
        - $ref: "#/components/parameters/CardId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuizSubmission" }
      responses:
        "200":
          description: 融合后的卡片
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          card: { $ref: "#/components/schemas/CardResponse" }
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /api/quiz/exam:
    post:
      tags: [quiz]
      summary: 智能组卷（跨知识点综合题）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExamRequest" }
      responses:
        "200":
          description: 试卷题目
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/ExamResult" }
        "400":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /api/quiz/mastery/{card_id}:
    get:
      tags: [quiz]
      summary: 获取单张卡片掌握度
      parameters:
        - $ref: "#/components/parameters/CardId"
      responses:
        "200":
          description: 掌握度
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [success] }
                  data: { $ref: "#/components/schemas/MasteryResult" }
  /api/quiz/mastery/batch:
    get:
      tags: [quiz]
      summary: 批量获取掌握度
      parameters:
        - name: card_ids
          in: query
          required: false
          schema: { type: string }
          description: 逗号分隔的卡片 ID 列表
      responses:
        "200":
          description: 批量掌握度
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          mastery:
                            type: object
                            additionalProperties: { type: integer }

  # ── Settings ───────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 当 STUDYWIKI_AUTH_TOKEN 配置时启用；public 路径无需携带。
  parameters:
    CardId:
      name: card_id
      in: path
      required: true
      schema: { type: string }
      description: 卡片 ID
...
  responses:
    Error:
      description: 统一错误响应
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
    ApiResponse:
      type: object
      description: 统一 API 响应格式
      required: [status]
      properties:
        status:
          type: string
          enum: [success, error]
        data:
          type: object
          nullable: true
        message:
          type: string
          default: ""
        error_code:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      description: 统一错误响应体
      required: [status, error_code, message, timestamp]
      properties:
        status:
          type: string
          enum: [error]
        error_code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    ErrorCode:
      type: string
      description: |
        统一错误码，格式 `SW-<DOMAIN>-<CODE>`。
        数值后缀通常对应该错误建议的 HTTP 状态码。
      enum:
        - SW-AUTH-001
        - SW-CARD-400
        - SW-CARD-404
        - SW-CARD-500
        - SW-UPLOAD-400
        - SW-UPLOAD-413
        - SW-UPLOAD-500
        - SW-QUIZ-400
        - SW-QUIZ-404
        - SW-QUIZ-500
        - SW-TASK-404
        - SW-SETTINGS-400
        - SW-KB-400
        - SW-KB-404
        - SW-KB-500
        - SW-BOOTSTRAP-400
        - SW-BOOTSTRAP-401
        - SW-BOOTSTRAP-429
        - SW-BOOTSTRAP-UPSTREAM
        - SW-BOOTSTRAP-TIMEOUT
        - SW-BOOTSTRAP-NETWORK
        - SW-AGENT-400
        - SW-AGENT-429
        - SW-LLM-503
        - SW-BACKUP-404
        - SW-BACKUP-500
        - SW-PAIR-400
        - SW-PAIR-401
        - SW-PAIR-404
        - SW-DB-507
        - SW-GENERIC-400
        - SW-GENERIC-401
        - SW-GENERIC-403
        - SW-GENERIC-404
        - SW-GENERIC-405
        - SW-GENERIC-409
        - SW-GENERIC-413
        - SW-GENERIC-422
        - SW-GENERIC-500
        - SW-GENERIC-502
        - SW-GENERIC-503
        - SW-GENERIC-504
...
    CardResponse:
      type: object
      description: 卡片响应
      required: [id, title, category, created_at, updated_at]
      properties:
        id: { type: string }
        title: { type: string }
        aliases:
          type: array
          items: { type: string }
        content: { type: string }
        examples:
          type: array
          items: { type: string }
        questions:
          type: array
          items: { type: string }
        category: { type: string }
        source_file: { type: string }
        source_page: { type: integer }
        related_cards:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    QuizAnswer:
      type: object
      properties:
        question: { type: string, default: "" }
        answer: { type: string, default: "" }
    QuizSubmission:
      type: object
      required: [card_id, answers]
      properties:
        card_id: { type: string }
        answers:
          type: array
          items: { $ref: "#/components/schemas/QuizAnswer" }
    QuizResult:
      type: object
      properties:
        question: { type: string }
        answer: { type: string }
        score:
          type: integer
          minimum: 0
          maximum: 10
        comment: { type: string }
        reference: { type: string, default: "" }
    QuizGenerateResult:
      type: object
      properties:
        card_id: { type: string }
        questions:
          type: array
          items:
            type: object
            properties:
              question: { type: string }
              ref_answer: { type: string }
    QuizGradeResult:
      type: object
      properties:
        card_id: { type: string }
        results:
          type: array
          items: { $ref: "#/components/schemas/QuizResult" }
        total_score: { type: integer }
        max_score: { type: integer }
        mastery_pct: { type: integer }
    ExamRequest:
      type: object
      required: [card_ids]
      properties:
        card_ids:
          type: array
          items: { type: string }
    ExamResult:
      type: object
      properties:
        questions:
          type: array
          items:
            type: object
            properties:
              question: { type: string }
              ref_answer: { type: string }
              related_cards:
                type: array
                items: { type: string }
        card_ids:
          type: array
          items: { type: string }
    MasteryResult:
      type: object
      properties:
        card_id: { type: string }
        attempts: { type: integer }
        score: { type: integer }
        mastery_pct: { type: integer }
...